Address PR review nits for PortScanner leak fix
