+++ cmuxTests/GhosttyConfigTests.swift
@@ -4000,34 +4000,3 @@ final class BrowserImportScopeTests: XCTestCase {
         XCTAssertNil(scope)
     }
 }
-
-final class PortScannerProcessCaptureTests: XCTestCase {
-    private func openFDCount() -> Int? {
-        try? FileManager.default.contentsOfDirectory(atPath: "/dev/fd").count
-    }
-
-    func testCaptureStandardOutputDoesNotLeakPipeFDs() throws {
-        guard let baseline = openFDCount() else {
-            throw XCTSkip("Unable to inspect /dev/fd on this runner")
-        }
-
-        var maxCount = baseline
-        for _ in 0..<200 {
-            let output = PortScanner.captureStandardOutput(
-                executablePath: "/usr/bin/printf",
-                arguments: ["cmux"]
-            )
-            XCTAssertEqual(output, "cmux")
-            if let current = openFDCount() {
-                maxCount = max(maxCount, current)
-            }
-        }
-
-        guard let finalCount = openFDCount() else {
-            throw XCTSkip("Unable to inspect final /dev/fd count on this runner")
-        }
-
-        XCTAssertLessThanOrEqual(maxCount - baseline, 8)
-        XCTAssertLessThanOrEqual(finalCount - baseline, 8)
-    }
-}
