+++ cmuxTests/PortScannerTests.swift
@@ -1,0 +1,38 @@
+import XCTest
+
+#if canImport(cmux_DEV)
+@testable import cmux_DEV
+#elseif canImport(cmux)
+@testable import cmux
+#endif
+
+final class PortScannerProcessCaptureTests: XCTestCase {
+    private func openFDCount() -> Int? {
+        try? FileManager.default.contentsOfDirectory(atPath: "/dev/fd").count
+    }
+
+    func testCaptureStandardOutputDoesNotLeakPipeFDs() throws {
+        guard let baseline = openFDCount() else {
+            throw XCTSkip("Unable to inspect /dev/fd on this runner")
+        }
+
+        var maxCount = baseline
+        for _ in 0..<200 {
+            let output = PortScanner.captureStandardOutput(
+                executablePath: "/usr/bin/printf",
+                arguments: ["cmux"]
+            )
+            XCTAssertEqual(output, "cmux")
+            if let current = openFDCount() {
+                maxCount = max(maxCount, current)
+            }
+        }
+
+        guard let finalCount = openFDCount() else {
+            throw XCTSkip("Unable to inspect final /dev/fd count on this runner")
+        }
+
+        XCTAssertLessThanOrEqual(maxCount - baseline, 8)
+        XCTAssertLessThanOrEqual(finalCount - baseline, 8)
+    }
+}
